Reflect review comments

+++ external/model-preparation-algorithm/mpa_tasks/utils/data_utils.py
@@ -42,9 +42,6 @@ def get_image(results: Dict[str, Any], cache_dir: str, to_float32=False):
     def is_training_subset(subset):
         return subset.name in ["TRAINING", "VALIDATION"]
 
-    def is_cached_or_caching(filename, tmp_filename):
-        return os.path.exists(filename) or os.path.exists(tmp_filename)
-
     def load_image_from_cache(filename: str, to_float32=False):
         try:
             cached_img = cv2.imread(filename)
@@ -55,7 +52,10 @@ def get_image(results: Dict[str, Any], cache_dir: str, to_float32=False):
             logger.warning(f"Skip loading cached {filename} \nError msg: {e}")
             return None
     
-    def save_image_to_cache(img: np.array, filename: str, tmp_filename: str):
+    def save_image_to_cache(img: np.array, filename: str):
+        tmp_filename = filename.replace(".png", "-tmp.png")
+        if os.path.exists(filename) or os.path.exists(tmp_filename):  # if image is cached or caching
+            return None
         try:
             cv2.imwrite(tmp_filename, img=img)
             os.replace(tmp_filename, filename)
@@ -67,7 +67,6 @@ def get_image(results: Dict[str, Any], cache_dir: str, to_float32=False):
     if is_training_subset(subset) and is_video_frame(results["dataset_item"].media):
         index = results["index"]
         filename = os.path.join(cache_dir, f"{subset}-{index:06d}.png")
-        tmp_filename = filename.replace(".png", "-tmp.png")
         if os.path.exists(filename):
             loaded_img = load_image_from_cache(filename, to_float32=to_float32)
             if loaded_img is not None:
@@ -77,7 +76,7 @@ def get_image(results: Dict[str, Any], cache_dir: str, to_float32=False):
     if to_float32:
         img = img.astype(np.float32)
 
-    if is_training_subset(subset) and is_video_frame(results["dataset_item"].media) and not is_cached_or_caching(filename, tmp_filename):
-        save_image_to_cache(img, filename, tmp_filename)
+    if is_training_subset(subset) and is_video_frame(results["dataset_item"].media):
+        save_image_to_cache(img, filename)
 
     return img
